ENH: Added functionality to Pyqtgraph for transform node
Implemented basic DateFormatter for Pyqtgraph tick lines.
Added TransformPlot class to handle creation of PyQtgraph Plot Widgets
for transform interface - sets plots with DateAxis and links the x-axis
scales.
Change Node Library initialization - add module parameter that
determines whether to present Nodes for use by the user, or to hide them
as is the case for the Display/Plot nodes - which require special
instantiation.
Updated FIRLowpassFilter to include the Series name in its output (for
use with plot legend)