Add Override classes for line selection click handling.
Add LinearFlightRegion subclass of PyQtGraphs LinearRegionItem which is
used to replicate the old flight-line selection behavior allowing users
to click and drag/resize rectangular areas on the plot.
Integrated this into plotters.py to create new LFR's on click when in
line-selection mode.
TODO: Figure out how/where the best way to place labels on the regions
is.